requirements.txt
asgiref==3.6.0
Django==4.1.6
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.0.0
Pillow==9.4.0
python-decouple==3.7
sqlparse==0.4.3
